requirements.txt
dnspython==2.6.1
Faker==25.0.1
psycopg2-binary==2.9.9
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
